restructured project to fit Next routing
