+++ frontend/src/app/(components)/Navbar.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { AppBar } from '@mui/material'
+const Navbar = () => {
+  return (
+    <div className='flex flex-col'>
+            <div className='flex-row space-x-10'>
+                <div className='left-5'>1</div>
+                <div>2</div>
+                <div>3</div>
+            </div>
+            <div>cwd/path</div>
+        </div>
+  )
+}
+
+export default Navbar
