+++ frontend/src/app/(components)/Sidebar.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Sidebar = () => {
+  return (
+    <aside>Sidebar</aside>
+  )
+}
+
+export default Sidebar
